Refacturing and Small code fixing
fixing:
Clients can now reconnect to the raspberry after a closed connection

Refacturing:
created the ObjConnClient. It is a Class for sending Objects to a server
and receiving Objects to a server. The ObjConnClient Class is used by
the RequestConnClient Class and it is planned that it will be used by
the unimplemented StatusConnClient Class