Change commands, unite behaviour for I²C devices and removal of obsolete newlines
The implemented commands now work upon the raspberrypi object as a
single interface instead of accessing it's members, as was decided in a
teammeeting on saturday. Furthermore removed some seemingly obsolete
methods (can be restored if necessary) and check the I²C behaviour for
usage on a "real emulated HI", which showed some necessary changes to
get everything to work. Finally this commit removes every encountered
obsolete newline and added an isInitialized() method for raspberry and
it's hardware components; The raspberry's isInitialized returns true
only if all components return true on their isInitialized(). This will
probably be extended or changed, but for now it reflects the desired
behaviour. This backend now has the following known issues:
* turning backlight on and off adds a weird symbol to the LCD
* GPIOpins read returns it's last *written* state, not the real state of
current flowing (as does GPIOpin.read(), but now we do not change to
Input to read, as this will effectively turn off pins if high before)
* due to the lack of a multimeter the potentiometer could not be tested