implemented RPC basis for the Registry Server and a simple test for it
The Registry server can accept Requests from the Network now
It listens on port 54320 for Requests

Incoming Requests will be handled by calling the appropriate method on
the server

In client connection a overloading Method for connectAsync was added,
which is used by the Test

The Test runs on RaspberryBackend and trys to call the register Method
over the network. The RegistryServer will call the appropriate method
and print "RegistryService.register called" in it. The test is no unit
test, because it looks almost identical to future working code.