RPC imporvements (see description)
Deleted the Request Handler, Abstract Command and Command Interface.
They got replaced by the Skeleton class. The Skeleton Class got moved
into the CommonFiles folder, because it can be used in the Raspberry
Backend and in the Registry Server. The main Source File of the
Raspberry Backend got cleaner as well because the Skeleton does the
server tasks.

Changed the Commands to Methods of the RaspberryPi class, which use
improved interfaces. EG ConnectPins(int x, int y) instead of
execute(Object parameter)

Komments still need to be adapted.