Feature Id 48: Implemented a possible c# API and improved networking
The API is the Raspberry Pi class in the Frontend. It can be used to remotely call operations and to register for notifications. It is only tested with the LightLED command, because the Frontend still has to use the API (thats why so much code is uncommented)

Improved Networking aswell: Receive Object got improved, so that Objects of different Types can be received over the same socket. This allowed to write the smaller SuccessResult and ExceptionResult classes, that replace the bigger Result class.